refactor: format args in editor
